+++ src/vec3/mod.rs
@@ -2,7 +2,7 @@ pub mod functions;
 
 #[derive(Debug, Clone, Copy, Default)]
 pub struct Vec3 {
-    e: [f32; 3],
+    e: [f64; 3],
 }
 
 pub type Color = Vec3;
